app/requirements.txt
aiogram>=3.0,<4.0
asyncpg>=0.29,<1.0
python-dotenv>=1.0,<2.0
aiohttp>=3.9,<4.0
